% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.cv.pogsnet}
\alias{plot.cv.pogsnet}
\title{Plot pogsnet CV}
\usage{
\method{plot}{cv.pogsnet}(x, sign.lambda, ...)
}
\arguments{
\item{x}{pogsnet.cv object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative if sign.lambda=-1.}

\item{...}{Other graphical parameters to plot.}
}
\description{
Plots results of pogsnet CV
}

