% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.pogsnet}
\alias{print.pogsnet}
\title{Prints pogsnet summary}
\usage{
\method{print}{pogsnet}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{pogsnet object}

\item{digits}{Number of digits to display}

\item{...}{Not used.}
}
\description{
Prings pogsnet summary
}

